# portal/public/openapi.yaml
openapi: 3.0.3
info:
  title: eGain Knowledge API
  version: "1.0.0"
  description: >
    REST API for knowledge search, articles CRUD, analytics, and agent-assisted suggestions.
    Versioning: base path /v1; backwards-incompatible changes ship under /v2. Backwards-compatible changes
    are additive and may be indicated via Deprecation/Sunset headers.
servers:
  - url: http://localhost:3000/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/token:
    post:
      tags: [Auth]
      summary: Get access token (demo/dev)
      description: >
        Issues a JWT for testing secured endpoints. For production, replace with OAuth2/OIDC.
      operationId: issueToken
      security: []   # <-- important: this endpoint itself is not secured
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [clientId, clientSecret]
              properties:
                clientId:
                  type: string
                  default: demo
                clientSecret:
                  type: string
                  format: password
                  default: demo-secret
            examples:
              demo:
                summary: Demo credentails
                value:
                  clientId: demo
                  clientSecret: demo-secret        
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [accessToken, tokenType, expiresIn]
                properties:
                  accessToken:
                    type: string
                  tokenType:
                    type: string
                    enum: [Bearer]
                    example: Bearer
                  expiresIn:
                    type: integer
                    description: Seconds until expiry
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search:
    post:
      tags: [Search]
      summary: Search knowledge articles
      description: >
        Full-text + metadata search across knowledge base with filters, sorting, and pagination.
        Rate limiting applies per-tenant and per-user. Responses include request correlation and rate limit headers.
      operationId: searchArticles
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/XTenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            examples:
              basic:
                summary: Search with filters
                value:
                  query: "password reset"
                  filters:
                    category: "Troubleshooting"
                    tags: ["SSO", "Login"]
                    fromDate: "2025-01-01T00:00:00Z"
                    toDate: "2025-12-31T23:59:59Z"
                  sortBy: relevance
                  pagination:
                    page: 1
                    pageSize: 20
      responses:
        "200":
          description: Search results
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT contains tenant and role claims. RBAC examples:
        viewer: search/read; editor: create/update; admin: delete + analytics admin.
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Client-provided request id for tracing (echoed in response).
      schema:
        type: string
        maxLength: 128
    XTenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: >
        Tenant identifier when operating in a multi-tenant gateway. If omitted, derived from JWT claims.
      schema:
        type: string
        maxLength: 128
  headers:
    XRequestId:
      description: Request correlation id.
      schema:
        type: string
    XRateLimitLimit:
      description: Request limit per window for this identity (tenant/user).
      schema:
        type: integer
        format: int64
    XRateLimitRemaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        format: int64
    XRateLimitReset:
      description: Unix timestamp (seconds) when the current window resets.
      schema:
        type: integer
        format: int64
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      additionalProperties: false
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 512
        filters:
          $ref: "#/components/schemas/SearchFilters"
        sortBy:
          type: string
          enum: [relevance, date, popularity]
          default: relevance
        pagination:
          $ref: "#/components/schemas/PageRequest"
    SearchFilters:
      type: object
      additionalProperties: false
      properties:
        category:
          type: string
          maxLength: 64
        fromDate:
          type: string
          format: date-time
        toDate:
          type: string
          format: date-time
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 64
    PageRequest:
      type: object
      additionalProperties: false
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
    SearchResponse:
      type: object
      required: [results, page, pageSize, total]
      additionalProperties: false
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    Article:
      type: object
      required: [id, title, content, category, tags, relevanceScore, createdDate, lastUpdated, viewCount]
      properties:
        id:
          type: string
        title:
          type: string
          maxLength: 256
        content:
          type: string
        category:
          type: string
          maxLength: 64
        tags:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 64
        relevanceScore:
          type: number
          format: float
        createdDate:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        viewCount:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, requestId]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad Request
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidQuery:
              value:
                error:
                  code: INVALID_ARGUMENT
                  message: "query must be at least 1 character"
                  requestId: "req_123"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    PayloadTooLarge:
      description: Payload Too Large
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too Many Requests
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/XRateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/XRateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/XRateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            throttled:
              value:
                error:
                  code: RATE_LIMITED
                  message: "Too many requests. Please retry later."
                  requestId: "req_456"
                  details:
                    scope: "tenant:user"
                    windowSeconds: 60
    InternalServerError:
      description: Internal Server Error
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"